+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 v0.1.0
 * Initial release
+
+v0.1.1
+* Code clean up and addes some unit tests
+
+v0.1.2
+* Added length to  encodeBigInt()